+++ sketch_oxide/src/statistics/mod.rs
@@ -8,10 +8,12 @@ mod ams;
 mod hll_joint;
 mod k_ary;
 mod moments;
+mod morris;
 mod p_stable;
 
 pub use ams::AmsSketch;
 pub use hll_joint::HllJointEstimator;
 pub use k_ary::KArySketch;
 pub use moments::MomentsSketch;
+pub use morris::MorrisCounter;
 pub use p_stable::PStableLpSketch;